Removes some dead code.
The dead code mainly is types that became unused after
the transition to nanopb and the state exchange protocol.

Some more of it is files that were not used anywhere,
or only used in a bogus way. It is better to re-introduce
this code at a time when it can be done in a meaningful way,
otherwise it is confusing to readers and a source of overhead
to maintainers.