Rename SensorReadings proto -> SensorsProto.
Preparation for reclaiming the name "SensorReadings" as a C++ class.

This is inconsistent with the names of the other proto messages, but
given that we're going to have this C++ class, I think it will be really
hard to remember which is which if one of them doesn't include the word
"proto".

GPC: rename-sensor-readings