Disable boot wizard & enable serial port for data
The piwiz.desktop file is responsible for the first-time boot setup wizard. Since we configure the pi with this script, its not a very useful popup that gets in the way. If deleting the file is too extreme, I think we can just move it elsewhere like the Downloads folder. Tested this locally on my pi.

The --serial-port flag disables the fake data being fed to the GUI and instead connects it to the rPi's UART pins. Now, the GUI and controller actually can communicate.