Add fio2 support to run_covent_tests.py.
You can run the covent tests all with air from the blower, all with
pressurized air/oxygen, or some with blower and some with o2.

Also don't turn off the ventilator between tests.  Edwin requested this
so the data captured by external sources looks cleaner.  The data
captured by this script is unaffected because in any case we wait a few
seconds before starting to record.

GPC: fio2-in-run-covent-tests